requirements.txt
fastapi
uvicorn
uiautomation
Pillow
pydantic
pydantic-settings
